fix(ui): drop visible kbd shortcut chips from header
The kbd chips that PR #89 was supposed to remove ended up in the
deployed merge — the revert force-push raced the merge and lost.
This PR re-applies the revert as a forward fix.

Removes:
- ``html.Kbd("⌥R")`` / ``html.Kbd("⌥P")`` chips beside the region +
  persona selectors
- ``.gp-header__shortcut-group`` wrappers + their aria-label
  containers
- ``.gp-header__kbd-hint`` CSS block + the ``body.briefing``
  override that hid them in projection mode

Keeps:
- The Alt+R / Alt+P / Alt+1..4 keyboard handlers in
  ``assets/accessibility.js`` (small surface, no maintenance cost,
  fine to leave for the rare power user even though they aren't
  advertised).
- The sticky header + tab strip from PR #89 — that's the actual
  win and is unchanged here.

## Why drop them

Visible shortcut chips work in apps where users live daily (Linear,
Vercel, GitHub) — every saved click adds up. GridPulse is a 5-tab
portfolio dashboard. Visitors are recruiters and interviewers
spending 90 seconds clicking through to evaluate the work. They
won't learn ``⌥R``. The chips signal "I noticed Linear has shortcut
hints" without solving any UX problem, and add chrome density to a
header that's deliberately minimalist.

## Tests

``tests/unit/test_sticky_chrome.py`` (renamed from
``test_sticky_chrome_kbd_hints.py``) — adds a
``TestKbdChipsAreNotShipped`` class that asserts:
- ``.gp-header__kbd-hint`` and ``.gp-header__shortcut-group``
  are absent from the CSS
- ``_build_header()`` produces zero elements with either class

These guards catch a future re-introduction so the call gets
reconsidered rather than slipping in silently.

All 1655 tests pass. Lint clean.
