+++ assets/custom.css
@@ -364,39 +364,6 @@ select:disabled,
     opacity: 0.6;
 }
 
-/* ── Keyboard-shortcut chips ──────────────────────────────────── */
-/* Sits adjacent to a selector to advertise its Alt+key shortcut
- * (handlers live in assets/accessibility.js). aria-hidden because
- * the underlying selector already carries aria-keyshortcuts —
- * screen-reader users don't need the visual chip read aloud. */
-.gp-header__kbd-hint {
-    display: inline-flex;
-    align-items: center;
-    margin-left: -4px;     /* tucked next to the preceding selector */
-    margin-right: var(--space-2);
-    pointer-events: none;  /* purely informational */
-}
-
-.gp-header__kbd-hint kbd {
-    font-family: var(--font-mono);
-    font-size: 10px;
-    font-weight: 500;
-    line-height: 1;
-    padding: 3px 5px;
-    border: 1px solid var(--border-default);
-    border-radius: var(--radius-sm);
-    background: var(--bg-elevated);
-    color: var(--text-tertiary);
-    box-shadow: 0 1px 0 var(--border-subtle);
-    letter-spacing: 0.02em;
-}
-
-/* Hide the kbd chips under Briefing Mode chrome — the projection
- * surface shouldn't show shortcut hints to a meeting audience. */
-body.briefing .gp-header__kbd-hint {
-    display: none;
-}
-
 /* ── Briefing Mode (R5c — projection / meeting-ready chrome) ───── */
 /*
  * Toggled via body.briefing (clientside mirror of meeting-mode-store).
